Finished Scale Degrees exercise.
To stop the chord progression before playing the answer note, I rewrote
stopAll() to return a Promise that resolves when all the notes have
stopped. I then play the answer note when this Promise resolves. At a
later date, I should rewrite the other references to stopAll() to take
advantage of this new functionality.